refactor: MVVM 패턴에 맞게 구조 재편
