+++ SnapPop/ViewModels/SnapComparison/SnapComparisonViewModel.swift
@@ -151,54 +151,4 @@ class SnapComparisonViewModel {
     func numberOfRows(in section: Int) -> Int {
         1
     }
-    
-    // MARK: - SnapComparisonSheetViewController Properties
-    var selectedIndex: Int = 0
-    var currentDateIndex: Int = 0
-    var currentPhotoIndex: Int = 0
-    var currentSnap: Snap? {
-        return filteredSnapData[currentDateIndex]
-    }
-    var isLeftArrowHidden: Bool {
-        return currentDateIndex == 0
-    }
-    var isRightArrowHidden: Bool {
-        return currentDateIndex == filteredSnapData.count - 1
-    }
-    
-    // 클로저
-    var updateUI: (() -> Void)?
-    var updatePageControl: ((Int, Int) -> Void)?
-    var updateArrowVisibility: ((Bool, Bool) -> Void)?
-    
-    // MARK: - SnapComparisonSheetViewController Methods
-    
-    func updateSnapData() {
-        updateUI?()
-        updatePageControl?(currentPhotoIndex, currentSnap?.images.count ?? 0)
-        updateArrowVisibility?(isLeftArrowHidden, isRightArrowHidden)
-    }
-    
-    func getSnapPhoto(at index: Int) -> UIImage? {
-        guard let currentSnap = currentSnap,
-              index >= 0,
-              index < currentSnap.images.count else {
-            return nil
-        }
-        return currentSnap.images[index]
-    }
-    
-    func moveToPreviousSnap() {
-        guard currentDateIndex > 0 else { return }
-        currentDateIndex -= 1
-        currentPhotoIndex = 0
-        updateSnapData()
-    }
-    
-    func moveToNextSnap() {
-        guard currentDateIndex < filteredSnapData.count - 1 else { return }
-        currentDateIndex += 1
-        currentPhotoIndex = 0
-        updateSnapData()
-    }
 }
